MusicRec/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aliyun-python-sdk-core=2.13.26=pypi_0
asgiref=3.3.0=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py37he774522_1000
ca-certificates=2020.7.22=0
certifi=2020.6.20=py37_0
cffi=1.14.3=py37h7a1dbc1_0
chardet=3.0.4=py37_1003
cryptography=3.1.1=py37h7a1dbc1_0
django=3.1.2=pypi_0
django-cors-headers=3.5.0=pypi_0
django-extensions=3.0.9=pypi_0
idna=2.10=py_0
intel-openmp=2020.2=254
jmespath=0.10.0=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.2.0=py37h45dec08_0
mkl_random=1.1.1=py37h47e9c7a_0
mysql-connector-c=6.1.11=h33f27b4_1
mysqlclient=2.0.1=pypi_0
numpy=1.18.1=pypi_0
numpy-base=1.19.1=py37ha3acd2a_0
openssl=1.1.1h=he774522_0
packaging=20.4=pypi_0
pip=20.2.3=py37_0
pip-review=1.1.0=pypi_0
pycparser=2.20=py_2
pycrypto=2.6.1=py37he774522_10
pymysql=0.10.1=py37_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pypi_0
pysocks=1.7.1=py37_1
python=3.7.0=hea74fb7_0
pytz=2020.1=py_0
requests=2.23.0=py37_0
setuptools=50.3.0=py37h9490d1a_1
six=1.15.0=py_0
sqlparse=0.4.1=py_0
urllib3=1.25.10=py_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_3
wheel=0.35.1=py_0
win_inet_pton=1.1.0=py37_0
wincertstore=0.2=py37_0
